# optimize encoder for common types