--- a/adapters/net/encode.go
+++ b/adapters/net/encode.go
@@ -9,25 +9,25 @@ import (
 
 func encodeTCPAddr(e objconv.Encoder, v reflect.Value) error {
 	a := v.Interface().(net.TCPAddr)
-	return e.Emitter.EmitString(a.String())
+	return e.Encode(a.String())
 }
 
 func encodeUDPAddr(e objconv.Encoder, v reflect.Value) error {
 	a := v.Interface().(net.UDPAddr)
-	return e.Emitter.EmitString(a.String())
+	return e.Encode(a.String())
 }
 
 func encodeUnixAddr(e objconv.Encoder, v reflect.Value) error {
 	a := v.Interface().(net.UnixAddr)
-	return e.Emitter.EmitString(a.String())
+	return e.Encode(a.String())
 }
 
 func encodeIPAddr(e objconv.Encoder, v reflect.Value) error {
 	a := v.Interface().(net.IPAddr)
-	return e.Emitter.EmitString(a.String())
+	return e.Encode(a.String())
 }
 
 func encodeIP(e objconv.Encoder, v reflect.Value) error {
 	a := v.Interface().(net.IP)
-	return e.Emitter.EmitString(a.String())
+	return e.Encode(a.String())
 }