This leads a small throughput increase on the encoder benchmark and enables a memory optimization on the application side:
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkCodeEncoder-4     12958944      12751068      -1.60%

benchmark                  old MB/s     new MB/s     speedup
BenchmarkCodeEncoder-4     149.74       152.18       1.02x

benchmark                  old allocs     new allocs     delta
BenchmarkCodeEncoder-4     0              0              +0.00%

benchmark                  old bytes     new bytes     delta
BenchmarkCodeEncoder-4     0             0             +0.00%
```