@achille-roussel 2017-06-12
@f2prateek are you satisfied with the README update?